extract suggestion, chip, loader components from searchbox and autocomplete
