+++ client/src/components/elements/AutocompleteLoader.js
@@ -6,7 +6,7 @@ import {
   parse as parseQueryString,
 } from 'query-string';
 
-export default class GeneAutocompleteLoader extends React.Component {
+export default class AutocompleteLoader extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -29,6 +29,19 @@ export default class GeneAutocompleteLoader extends React.Component {
       this.loadSuggestions(this.props.inputValue);
     }
   }
+
+  getSuggestionFromMatches = (match) => {
+    return Object.keys(match).reduce((result, key) => {
+      const value = match[key];
+      const [namespace, keyName] = key.split('/');
+      let newPairs = keyName === 'id' ? { entityType: namespace } : {};
+      return {
+        ...result,
+        [keyName]: value,
+        ...newPairs,
+      };
+    }, {});
+  };
 
   loadSuggestions = (inputValue, isInitialLoad) => {
     if (inputValue.length < 2) {
@@ -65,7 +78,9 @@ export default class GeneAutocompleteLoader extends React.Component {
             );
           },
           () => {
-            return fetch(`/api/gene/?pattern=${inputValue}`);
+            return fetch(
+              `/api/${this.props.entityType}/?pattern=${inputValue}`
+            );
           }
         )
           .then((response) => {
@@ -78,14 +93,8 @@ export default class GeneAutocompleteLoader extends React.Component {
             if (matchedPattern === this.state.inputValue) {
               // to avoid problem caused by response coming back in the wrong order
               // compare inputValue to produce suggestion with current inputValue,
-              const { matches } = content;
-              const suggestions = matches.map((item) => ({
-                id: item['gene/id'],
-                label:
-                  item['gene/cgc-name'] ||
-                  item['gene/sequence-name'] ||
-                  item['gene/id'],
-              }));
+              const { matches = [] } = content;
+              const suggestions = matches.map(this.getSuggestionFromMatches);
               this.setState(
                 {
                   suggestions,
@@ -109,7 +118,8 @@ export default class GeneAutocompleteLoader extends React.Component {
   }
 }
 
-GeneAutocompleteLoader.propTypes = {
+AutocompleteLoader.propTypes = {
+  entityType: PropTypes.string.isRequired,
   inputValue: PropTypes.string,
   selectedValue: PropTypes.string,
   onSuggestionChange: PropTypes.func,
