+++ client/src/containers/Gene/GeneAutocomplete.js
@@ -2,15 +2,16 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import {
   AutocompleteBase,
+  AutocompleteChip,
+  AutocompleteLoader,
+  AutocompleteSuggestion,
   withStyles,
-  Chip,
   InputAdornment,
   Paper,
   MenuItem,
   TextField,
   SimpleListPagination,
 } from '../../components/elements';
-import GeneAutocompleteLoader from './GeneAutocompleteLoader';
 
 function renderInput(inputProps) {
   const { InputProps, classes, ref, item, reset, ...other } = inputProps;
@@ -26,12 +27,7 @@ function renderInput(inputProps) {
         },
         startAdornment: item ? (
           <InputAdornment position="start">
-            <Chip
-              tabIndex={-1}
-              label={`${item.label} [ID: ${item.id}]`}
-              className={classes.chip}
-              onDelete={reset}
-            />
+            <AutocompleteChip suggestion={item} onDelete={reset} />
           </InputAdornment>
         ) : null,
       }}
@@ -39,41 +35,6 @@ function renderInput(inputProps) {
     />
   );
 }
-
-function renderSuggestion({
-  suggestion,
-  index,
-  itemProps,
-  highlightedIndex,
-  selectedItem,
-}) {
-  const isHighlighted = highlightedIndex === index;
-  const isSelected = selectedItem === suggestion;
-
-  return (
-    <MenuItem
-      {...itemProps}
-      key={suggestion.label}
-      selected={isHighlighted}
-      component={'div'}
-      style={{
-        fontWeight: isSelected ? 500 : 400,
-      }}
-    >
-      {suggestion.label}
-    </MenuItem>
-  );
-}
-renderSuggestion.propTypes = {
-  highlightedIndex: PropTypes.number,
-  index: PropTypes.number,
-  itemProps: PropTypes.object,
-  selectedItem: PropTypes.string,
-  suggestion: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    label: PropTypes.string,
-  }).isRequired,
-};
 
 class GeneAutocomplete extends Component {
   itemToString = (item) => (item ? item.id : '');
@@ -103,7 +64,8 @@ class GeneAutocomplete extends Component {
           ...downshift
         }) => (
           <div>
-            <GeneAutocompleteLoader
+            <AutocompleteLoader
+              entityType="gene"
               inputValue={inputValue}
               selectedValue={this.itemToString(selectedItem)}
               onSuggestionChange={(suggestions, isInitialLoad) => {
@@ -144,15 +106,15 @@ class GeneAutocomplete extends Component {
                     {({ pageItems, navigation }) =>
                       isOpen ? (
                         <Paper className={classes.paper} square>
-                          {pageItems.map((suggestion, index) =>
-                            renderSuggestion({
-                              suggestion,
-                              index,
-                              itemProps: getItemProps({ item: suggestion }),
-                              highlightedIndex,
-                              selectedItem,
-                            })
-                          )}
+                          {pageItems.map((suggestion, index) => (
+                            <AutocompleteSuggestion
+                              suggestion={suggestion}
+                              index={index}
+                              highlightedIndex={highlightedIndex}
+                              selectedItem={selectedItem}
+                              itemProps={getItemProps({ item: suggestion })}
+                            />
+                          ))}
                           {navigation}
                         </Paper>
                       ) : null
@@ -160,7 +122,7 @@ class GeneAutocomplete extends Component {
                   </SimpleListPagination>
                 </div>
               )}
-            </GeneAutocompleteLoader>
+            </AutocompleteLoader>
           </div>
         )}
       </AutocompleteBase>
